+++ handlers.ts
@@ -1,30 +1,29 @@
-import * as AWS from 'aws-sdk';
 import express, { Request, Response } from 'express';
 import serverless from 'serverless-http';
-import { Octokit } from "@octokit/rest";
-import { createAppAuth } from "@octokit/auth-app";
+import AWS from 'aws-sdk'; // Ensure AWS SDK supports ES module import
+import { Octokit } from '@octokit/rest';
+import { createAppAuth } from '@octokit/auth-app';
 import { getSecret } from './secrets';
+import { getUser } from './users';
+import { getProjectData, storeProjectData, SourceType, convertToSourceType } from './storage';
+import { validateUser } from './auth';
 
 const app = express();
-const dynamoDB = new AWS.DynamoDB.DocumentClient();
 
-const installationsKeyValueStore = 'Boost.GitHub-App.installations';
 const BoostGitHubAppId = "472802";
 
-app.use(express.json()); // Use express.json middleware to parse JSON request body
+app.use(express.json()); // Make sure to use express.json middleware to parse JSON request body
 
 app.get('/api/get_file_from_uri', async (req: Request, res: Response) => {
+    const email = validateUser(req, res);
+    if (!email) {
+        return;
+    }
+
     if (!req.query.uri) {
         console.error(`URI is required`);
         return res.status(400).send('URI is required');
     }
-
-    if (!req.query.email) {
-        console.error(`Unauthorized:  Email is required`);
-        return res.status(401).send('Unauthorized');
-    }
-
-    const email = normalizeEmail(req.query.email as string);
 
     const uri = new URL(req.query.uri as string);
     if (uri.protocol !== 'http:' && uri.protocol !== 'https:') {
@@ -44,37 +43,36 @@ app.get('/api/get_file_from_uri', async (req: Request, res: Response) => {
 
     console.log(`Inboumd Request: ${JSON.stringify(payload)}`);
 
-    let installationId: string;
-    try {
-        const params = {
-            TableName: installationsKeyValueStore,
-            Key: { email }
-        };
-
-        const userInfo = await dynamoDB.get(params).promise();
-        installationId = userInfo.Item?.installationId; // Add null check here
-        const installingUser = userInfo.Item?.username; // Add null check here
-
-    } catch (error) {
-        console.error(`Error retrieving installation user info:`, error);
+    const user = await getUser(email);
+    const installationId = user?.installationId;
+    if (!installationId) {
         return res.status(401).send('Unauthorized');
     }
 
+    // try to get the file from GitHub via public path without authentication
     try {
         const octokit = new Octokit();
         const response = await octokit.rest.repos.getContent({
-            owner,
-            repo,
+            owner: owner,
+            repo: repo,
             path: filePathWithoutBranch
         });
 
+        let fileContent = '';
+
         // Check if response is for a single file and has content
         if ("content" in response.data && typeof response.data.content === 'string') {
-            const fileContent = Buffer.from(response.data.content, 'base64').toString('utf8');
-            return res.send(fileContent);
+            fileContent = Buffer.from(response.data.content, 'base64').toString('utf8');
         } else {
             throw new Error('Content not found or not a file');
         }
+
+        // Set the custom header
+        // Example: 'X-Resource-Access' or public or private
+        const fileVisibility = 'public';
+        res.set('X-Resource-Access', fileVisibility);
+        
+        return res.send(fileContent);
 
     } catch (error: any) {
         if (error.status !== 404) {
@@ -82,49 +80,140 @@ app.get('/api/get_file_from_uri', async (req: Request, res: Response) => {
         } else if (error?.response?.data?.message === 'Not Found') {
             console.error(`Failed to retrieve file via public access`);
         } else {
-            console.error(`Error: retrieving file via public access`, error);
+        console.error(`Error: retrieving file via public access`, error);
         }
     }
 
     try {
+
         const secretStore = 'boost/GitHubApp';
         const secretKeyPrivateKey = secretStore + '/' + 'private-key';
 
         const privateKey = await getSecret(secretKeyPrivateKey);
 
+        // Configure the auth strategy for Octokit
+        const auth = createAppAuth({
+            appId: BoostGitHubAppId,
+            privateKey: privateKey,
+            installationId: installationId,
+        });
+
         const octokit = new Octokit({
             authStrategy: createAppAuth,
             auth: {
                 appId: BoostGitHubAppId,
-                privateKey,
-                installationId
+                privateKey: privateKey,
+                installationId: installationId,
             }
         });
 
         const response = await octokit.rest.repos.getContent({
-            owner,
-            repo,
+            owner: owner,
+            repo: repo,
             path: filePathWithoutBranch
         });
 
+        let fileContent = '';
         // Check if response is for a single file and has content
         if ("content" in response.data && typeof response.data.content === 'string') {
-            const fileContent = Buffer.from(response.data.content, 'base64').toString('utf8');
+            fileContent = Buffer.from(response.data.content, 'base64').toString('utf8');
             console.log(`File returned: Owner: ${owner}, Repo: ${repo}, Path: ${filePathWithoutBranch}`);
-            return res.send(fileContent);
         } else {
             throw new Error('Content not found or not a file');
         }
 
-    } catch (error: any) {
+        // Set the custom header
+        // Example: 'X-Resource-Access' or public or private
+        const fileVisibility = 'private';
+        res.set('X-Resource-Access', fileVisibility);
+
+        return res.send(fileContent);
+        
+    } catch (error) {
         console.error(`Error:`, error);
         return res.status(500).send('Internal Server Error');
     }
 });
 
-function normalizeEmail(email: string): string {
-    email = email.toLowerCase();
-    return email.replace(/@polytest\.ai$/i, '@polyverse.com');
-}
+app.get('/api/files/:source/:owner/:project/:pathBase64/:analysisType', async (req, res) => {
+    try {
+        const email = validateUser(req, res);
+        if (!email) {
+            return;
+        }
 
-export const getFromFileURI = serverless(app);
+        const { source, owner, project, pathBase64, analysisType } = req.params;
+
+        if (!source || !owner || !project || !pathBase64 || !analysisType) {
+            if (!source) {
+                console.error(`Source is required`);
+            } else if (!owner) {
+                console.error(`Owner is required`);
+            } else if (!project) {
+                console.error(`Project is required`);
+            }
+            else if (!pathBase64) {
+                console.error(`Path is required`);
+            }
+            else if (!analysisType) {
+                console.error(`Analysis type is required`);
+            }
+            return res.status(400).send('Invalid resource path');
+        }
+
+        let decodedPath;
+        try {
+            decodedPath = Buffer.from(pathBase64, 'base64').toString('utf8');
+        } catch (error) {
+            console.error(`Error decoding path: ${pathBase64}`, error);
+            return res.status(400).send(`Invalid resource path`);
+        }
+
+        const data = await getProjectData(email, convertToSourceType(source), owner, project, decodedPath, analysisType);
+        if (!data) {
+            console.error(`Resource not found: ${source}, ${owner}, ${project}, ${decodedPath}, ${analysisType}`);
+            return res.status(404).send('Resource not found');
+        }
+    } catch (error) {
+        console.error(`Handler Error: /api/files/:source/:owner/:project/:pathBase64/:analysisType`, error);
+        return res.status(500).send('Internal Server Error');
+    }
+});
+
+app.post('/api/files/:source/:owner/:project/:pathBase64/:analysisType', async (req, res) => {
+    try {
+        const email = validateUser(req, res);
+        if (!email) {
+            return res.status(401).send('Unauthorized');
+        }
+
+        const { source, owner, project, pathBase64, analysisType } = req.params;
+
+        if (!source || !owner || !project || !pathBase64 || !analysisType) {
+            console.error('Missing required parameters in request:', req.params);
+            return res.status(400).send('Invalid resource path');
+        }
+
+        let decodedPath;
+        try {
+            decodedPath = Buffer.from(pathBase64, 'base64').toString('utf8');
+        } catch (error) {
+            console.error(`Error decoding path: ${pathBase64}`, error);
+            return res.status(400).send('Invalid resource path');
+        }
+
+        const data = req.body; // Assuming data is sent directly in the body
+        if (!data) {
+            return res.status(400).send('No data provided');
+        }
+
+        await storeProjectData(email, convertToSourceType(source), owner, project, decodedPath, analysisType, data);
+        res.sendStatus(200);
+
+    } catch (error) {
+        console.error(`Handler Error: /api/files/:source/:owner/:project/:pathBase64/:analysisType`, error);
+        return res.status(500).send('Internal Server Error');
+    }
+});
+
+module.exports.getFromFileURI = serverless(app);